Separate the blueprint container into individual class.
